Use long to store ids.

+++ app/src/main/java/pl/arturborowy/rnm/data/remote/stations/keywords/KeywordDto.kt
@@ -4,9 +4,9 @@ import com.google.gson.annotations.SerializedName
 
 data class KeywordDto(
     @SerializedName("id")
-    val id: Int,
+    val id: Long,
     @SerializedName("keyword")
     val keyword: String,
     @SerializedName("station_id")
-    val stationId: Int
+    val stationId: Long
 )
